requirements.txt
amqp==5.1.1
asgiref==3.5.2
attrs==22.1.0
billiard==3.6.4.0
bleach==5.0.1
boto3==1.26.22
botocore==1.29.22
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.0.4
gunicorn==20.1.0
idna==3.4
iniconfig==1.1.1
install==1.3.5
jmespath==1.0.1
kombu==5.2.4
Markdown==3.3.4
martor==1.6.15
packaging==21.3
Pillow==9.3.0
pluggy==1.0.0
prompt-toolkit==3.0.33
# pycurl==7.45.1
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
